Refactor test base classes and remove DefaultMapperTest
Replaced DefaultMapperTest with AbstractMapperTest as the base class in InterceptingExecutorInterceptorTest and related tests. Moved test logic from DefaultMapperTest to InterceptingExecutorInterceptorTest, deleted DefaultMapperTest, and updated method visibility in AbstractSqlSessionTest. Renamed DefaultMyBatisTest to AbstractMapperTestTest to reflect the new base class.
